Aws Hackathon/AWS-Hackathon/google-reviews-scraper-pro/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
aiohttp==3.11.11
googletrans==4.0.2
selenium==4.15.2
undetected-chromedriver==3.5.4
tqdm==4.66.3
pymongo==4.12.0
pyyaml==6.0.1
certifi==2024.7.4
webdriver-manager==4.0.2
setuptools==79.0.1
boto3==1.35.1
pytest==7.4.3
pandas
pytz
fake-useragent
Flask
flask-cors==4.0.0

